# api/docs/swagger.yaml
definitions:
...
  models.CreatePositionModel:
    properties:
      company_id:
        type: string
      name:
        type: string
      position_attributes:
        items:
          $ref: '#/definitions/models.PositionAttribute'
        type: array
      profession_id:
        type: string
    type: object
...
  models.PositionAttribute:
    properties:
      attribute_id:
        type: string
      value:
        type: string
    type: object
...
  models.ResponseModel:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
...
  position_service.Attribute:
    properties:
      attribute_type:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  position_service.CreatePositionRequest:
    properties:
      company_id:
        type: string
      name:
        type: string
      position_attributes:
        items:
          $ref: '#/definitions/position_service.PositionAttribute'
        type: array
      profession_id:
        type: string
    type: object
...
  position_service.GetAllPositionResponse:
    properties:
      count:
        type: integer
      positions:
        items:
          $ref: '#/definitions/position_service.GetPositionResponse'
        type: array
    type: object
...
  position_service.GetPositionAttribute:
    properties:
      attribute:
        $ref: '#/definitions/position_service.Attribute'
      attribute_id:
        type: string
      id:
        type: string
      value:
        type: string
    type: object
  position_service.GetPositionResponse:
    properties:
      company_id:
        type: string
      id:
        type: string
      name:
        type: string
      position_attributes:
        items:
          $ref: '#/definitions/position_service.GetPositionAttribute'
        type: array
      profession_id:
        type: string
    type: object
  position_service.PositionAttribute:
    properties:
      attribute_id:
        type: string
      value:
        type: string
    type: object
  position_service.PositionStatus:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/position:
    get:
      consumes:
      - application/json
      description: get all position
      operationId: get-position-all
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: profession_id
        in: query
        name: profession_id
        type: string
      - description: company_id
        in: query
        name: company_id
        type: string
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/position_service.GetAllPositionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get position all
      tags:
      - position
    post:
      consumes:
      - application/json
      description: Create Position
      operationId: create-position
      parameters:
      - description: position
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/models.CreatePositionModel'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Create Position
      tags:
      - position
  /v1/position/{position_id}:
    delete:
      consumes:
      - application/json
      description: Delete Position by given ID
      operationId: delete-position
      parameters:
      - description: position_id
        in: path
        name: position_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/position_service.PositionStatus'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Delete Position
      tags:
      - position
    get:
      consumes:
      - application/json
      description: Get Position
      operationId: get-position-id
      parameters:
      - description: position_id
        in: path
        name: position_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/position_service.GetPositionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Get Position
      tags:
      - position
    put:
      consumes:
      - application/json
      description: Update Position by ID
      operationId: update-position
      parameters:
      - description: position_id
        in: path
        name: position_id
        required: true
        type: string
      - description: position
        in: body
        name: position
        required: true
        schema:
          $ref: '#/definitions/position_service.CreatePositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/position_service.PositionStatus'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Update Position
      tags:
      - position
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"